move enums to domain package
