+++ src/main/java/com/koc/user/domain/User.java
@@ -1,8 +1,6 @@
 package com.koc.user.domain;
 
 import com.koc.user.application.dto.UserDto;
-import com.koc.user.application.enums.LoginType;
-import com.koc.user.application.enums.UserStatus;
 import com.koc.user.adapter.out.persistence.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
